+++ app/Http/Controllers/AdminPaymentController.php
@@ -19,6 +19,9 @@ class AdminPaymentController extends Controller
      */
     public function index()
     {
+        if(Auth::user()->role_id != 1) {
+            return redirect('dashboard'); 
+        }
         $team = User::select('id')->whereIn('role_id', [2,3,4])->get()->toArray();
         $payment = PaymentConfirmation::get();
         return View('admin.payment.index')->with('payment', $payment);
@@ -87,6 +90,9 @@ class AdminPaymentController extends Controller
      */
     public function destroy($id)
     {
+        if(Auth::user()->role_id != 1) {
+            return redirect('dashboard'); 
+        }
         $payment = PaymentConfirmation::find($id);
         $payment->delete();
 
@@ -96,6 +102,9 @@ class AdminPaymentController extends Controller
 
     public function confirm($id)
     {
+        if(Auth::user()->role_id != 1) {
+            return redirect('dashboard'); 
+        }
         $user = User::find(PaymentConfirmation::find($id)->user_id);
         if($user->role_id == 4) {
             $user->team->hash = crypt($user->name . $user->created_at, 'vocomfest2017');
